feat: Rewrite this with serverless for node@8, borrowing from https://github.com/upsidetravel/bucket-antivirus-function.
One of the last of my TODOs from Fetch Auto. I mean, I'm not gonna use this I don't think, but I don't really want my name associated with something so old and terribly hacked at as that V1 from 2017.

Ideally we'd package the binaries up as a layer (so we don't upload them twice for both `scanFile` and `updateDefinitions`), but that can wait.